requirements.txt
fastapi==0.81.0
uvicorn==0.18.3
scikit-learn==1.0.2
joblib
pandas
numpy
mlserver
requests
Pillow
regex